cleanup Constructor code
